add simple distance and edge calc

+++ robot/src/game/movement.rs
@@ -2,13 +2,13 @@
 pub struct Move{
     pub row:usize,
     pub col:usize,
-    pub value:f32,
+    pub value:i32,
 }
 
 impl Move {
     pub fn new(row:usize, col:usize) -> Self{
         Self{
-            row, col, value:0.,
+            row, col, value:0,
         }
     }
 }
